feat(payment): 결제 승ìã인 서비ì스 구현

+++ payment/src/main/java/com/devticket/payment/payment/application/dto/PgPaymentConfirmCommand.java
@@ -3,6 +3,6 @@ package com.devticket.payment.payment.application.dto;
 public record PgPaymentConfirmCommand(
     String paymentKey,
     String orderId,
-    Long amount
+    Integer amount
 ) {
 }
