+++ payment/src/main/java/com/devticket/payment/payment/infrastructure/external/dto/TossPaymentConfirmRequest.java
@@ -3,6 +3,6 @@ package com.devticket.payment.payment.infrastructure.external.dto;
 public record TossPaymentConfirmRequest(
     String paymentKey,
     String orderId,
-    Long amount
+    Integer amount
 ) {
 }
